Incomplete Input

+++ Source/Engine/Core/Vector2.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <cmath>
 
 namespace kiko
 {
@@ -30,7 +31,7 @@ namespace kiko
 		Vector2& operator *= (const Vector2& v) { x *= v.x; y *= v.y; return *this; }
 
 		float LengthSqr() {	return (x * x) + (y * y); }
-		float Length() { return std::sqrt(LengthSqr()); }
+		float Length() { return static_cast<float>(std::sqrt(LengthSqr())); }
 
 		Vector2 Normalized() { return *this / Length(); }
 		void Normalize() { *this /= Length(); }
